backend/requirements.txt
python-dotenv
openai
yt-dlp
pydub
